api/requirements.txt
flask==3.1.0
flask-cors==5.0.0
pandas==2.2.3
google-generativeai==0.8.3
supabase==2.3.4
python-dotenv==1.0.0
